forge/pyproject.toml
[project]
name = "axm-forge"
version = "2.0.0"
description = "AXM Forge - ingestion + claim extraction + Clarion transport"
requires-python = ">=3.10"
dependencies = [
  "graphkdf>=1.0.0",
  "cryptography>=41.0.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
full = [
  "duckdb>=0.9.0",
]

[project.scripts]
axm-forge = "axm_forge.cli.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
